refactor: extract track utils
